pyproject.toml
[project]
name = "agentic-backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "alembic>=1.17.0",
    "asyncpg>=0.30.0",
    "authlib>=1.6.5",
    "bcrypt>=5.0.0",
    "fastapi[standard]>=0.118.0",
    "httpx>=0.28.1",
    "itsdangerous>=2.2.0",
    "passlib[bcrypt]>=1.7.4",
    "psycopg2-binary>=2.9.10",
    "pydantic-settings>=2.11.0",
    "pyjwt>=2.10.1",
    "python-dotenv>=1.1.1",
    "python-jose[cryptography]>=3.5.0",
    "requests>=2.32.5",
    "sqlalchemy>=2.0.43",
    "sqlmodel>=0.0.25",
    "uvicorn>=0.37.0",
]

[dependency-groups]
dev = [
    "anyio>=4.11.0",
    "httpx>=0.28.1",
    "pytest>=8.4.2",
    "pytest-asyncio>=1.2.0",
]
